+++ client/src/actions/bloodStore.js
@@ -86,18 +86,20 @@ export const updateBloodStore = (formData) => async (dispatch) => {
 
 // Delete blood store
 export const deleteBloodStore = (id) => async (dispatch) => {
-  try {
-    const res = await axios.delete(`/api/blood-store/${id}`);
-    dispatch({
-      type: UPDATE_BLOOD_STORE,
-      payload: res.data,
-    });
+  if (window.confirm('Are you sure? This can NOT be undone!')) {
+    try {
+      const res = await axios.delete(`/api/blood-store/${id}`);
+      dispatch({
+        type: UPDATE_BLOOD_STORE,
+        payload: res.data,
+      });
 
-    dispatch(setAlert('Successfully Deleted Blood Store', 'success'));
-  } catch (err) {
-    dispatch({
-      type: BLOOD_STORE_ERROR,
-      payload: { msg: err.response.statusText, status: err.response.status },
-    });
+      dispatch(setAlert('Successfully Deleted Blood Store', 'success'));
+    } catch (err) {
+      dispatch({
+        type: BLOOD_STORE_ERROR,
+        payload: { msg: err.response.statusText, status: err.response.status },
+      });
+    }
   }
 };
